Move the raw data out of public.
We don't want the users to view the raw CSV data. If the
file is in the public folder they'll be able to access it
by visiting the /season_games.csv path. Moving it to
another folder (in this case "data") prevents Sinatra from
serving it up to the user.